Get Property Api Implementation 2

+++ app/src/main/java/com/hotel/theconvo/data/remote/dto/response/GetPropertyResponse.kt
@@ -55,7 +55,7 @@ data class Property(
 )
 
 data class Proximity(
-    val km: Int,
+    val km: Double,
     val location: String
 )
